Many updates related to writing NBT tags to file
* Changed write_tags method parameters to accept a 'moved' integer pointer,
  which is used for recursion in the same way as read_tags works.
* Writing tags operates on a list of tag names and their known tag ID's, and
  adding bad tag names to the dictionary shouldn't harm it or result in them
  being written to file.  However, putting bad data in a given tag or the wrong
  type might.
* Updated World_save to match the new parameters, and added debugging output.

There are a still things that still need to be added, including writing a root
tag to the file, closing that root tag, and making sure the endian-ness is
correct (right now, I'll 99% sure it isn't).  Also, the list tag type needs to
be added; this may result in some refactoring.